+++ lib/observations/src/observation_value_view.dart
@@ -1,66 +1,14 @@
 import 'dart:collection';
 
-import 'package:fhir/r4/r4.dart';
+import 'package:fhir/r4.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 
-import '../../fhir_types/fhir_types.dart';
-
-/// Display a single [Observation] with value, code, and timestamp.
-/// Currently only supports observations with valueQuantity,
-/// or components of valueQuantity.
-class ObservationWidget extends StatelessWidget {
-  final Widget _valueWidget;
-  final Widget _codeWidget;
-  final Widget _dateTimeWidget;
-  final Locale? locale;
-
-  ObservationWidget(
-    Observation observation, {
-    Key? key,
-    this.locale,
-    TextStyle? valueStyle,
-    TextStyle? unitStyle,
-    TextStyle? codeStyle,
-    TextStyle? dateTimeStyle,
-    String componentSeparator = ' | ',
-    String unknownUnitText = '',
-    String unknownValueText = '?',
-  })  : _valueWidget = ObservationValueWidget(
-          observation,
-          valueStyle: valueStyle,
-          unitStyle: unitStyle,
-          locale: locale,
-          componentSeparator: componentSeparator,
-          unknownUnitText: unknownUnitText,
-          unknownValueText: unknownValueText,
-        ),
-        _codeWidget =
-            CodeableConceptText(observation.code, style: codeStyle, key: key),
-        _dateTimeWidget = FhirDateTimeText(
-          observation.effectiveDateTime,
-          locale: locale,
-          style: dateTimeStyle,
-        ),
-        super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        _valueWidget,
-        _codeWidget,
-        _dateTimeWidget,
-      ],
-    );
-  }
-}
-
 /// Display the value of a single [Observation].
+///
 /// Currently only supports observations with valueQuantity,
 /// or components of valueQuantity.
-class ObservationValueWidget extends StatelessWidget {
+class ObservationValueView extends StatelessWidget {
   final Observation _observation;
   final Locale? locale;
   final TextStyle? valueStyle;
@@ -69,7 +17,7 @@ class ObservationValueWidget extends StatelessWidget {
   final String unknownValueText;
   final String unknownUnitText;
 
-  const ObservationValueWidget(
+  const ObservationValueView(
     this._observation, {
     Key? key,
     this.locale,